+++ packages/frontend/src/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./queries";
 export { usePlayCooldownExpirySound, usePlaySound } from "./sfx";
 export { useCanvasImage } from "./useCanvasImage";
 export { useCanvasSearchParams } from "./useCanvasSearchParams";
+export { useIsFullscreenAvailable } from "./useIsFullscreenAvailable";
